+++ packages/ast/template-tag/tests/preprocess/class-2.test.ts
@@ -5,35 +5,17 @@ import { preprocess } from '../../src/index.js';
 test('preprocess > class (2)', function () {
   const oldFile = [
     `import Component from '@glimmer/component';`,
-    `import { local } from 'embroider-css-modules';`,
     ``,
-    `import styles from './example.css';`,
-    ``,
-    `const UserName = <template>`,
-    `  <div`,
-    `    class={{local styles "container" "highlight"}}`,
-    `    data-test-field="😀😀a🎉🎉"`,
-    `  >`,
-    `    {{@user.name}}`,
-    `  </div>`,
-    `</template>;`,
+    `import styles from './my-component.css';`,
     ``,
     `export default class MyComponent extends Component {`,
-    `  get timestamp(): string {`,
-    `    return 'yesterday';`,
-    `  }`,
-    ``,
     `  <template>`,
-    `    <UserName @user={{@user}} />`,
-    ``,
-    `    <div class={{styles.container}} data-test-field="timestamp">`,
-    `      {{this.timestamp}}`,
-    ``,
-    `  <p> 😀😀😀 Hello! 🎉🎉🎉`,
-    `  </p>`,
+    `    <div class={{styles.container}}>`,
+    `      Hello world!`,
     `    </div>`,
     `  </template>`,
     `}`,
+    ``,
   ].join('\n');
 
   const { javascript, templateTags } = preprocess(oldFile);
@@ -43,33 +25,12 @@ test('preprocess > class (2)', function () {
     [
       `import { template as template_fd9b2463e5f141cfb5666b64daa1f11a } from "@ember/template-compiler";`,
       `import Component from '@glimmer/component';`,
-      `import { local } from 'embroider-css-modules';`,
-      `import styles from './example.css';`,
-      `const UserName = template_fd9b2463e5f141cfb5666b64daa1f11a(\``,
-      `  <div`,
-      `    class={{local styles "container" "highlight"}}`,
-      `    data-test-field="😀😀a🎉🎉"`,
-      `  >`,
-      `    {{@user.name}}`,
-      `  </div>`,
-      `\`, {`,
-      `    eval () {`,
-      `        return eval(arguments[0]);`,
-      `    }`,
-      `});`,
+      `import styles from './my-component.css';`,
       `export default class MyComponent extends Component {`,
-      `    get timestamp(): string {`,
-      `        return 'yesterday';`,
-      `    }`,
       `    static{`,
       `        template_fd9b2463e5f141cfb5666b64daa1f11a(\``,
-      `    <UserName @user={{@user}} />`,
-      ``,
-      `    <div class={{styles.container}} data-test-field="timestamp">`,
-      `      {{this.timestamp}}`,
-      ``,
-      `  <p> 😀😀😀 Hello! 🎉🎉🎉`,
-      `  </p>`,
+      `    <div class={{styles.container}}>`,
+      `      Hello world!`,
       `    </div>`,
       `  \`, {`,
       `            component: this,`,
@@ -86,75 +47,30 @@ test('preprocess > class (2)', function () {
   assert.deepStrictEqual(templateTags, [
     {
       contentRange: {
-        endByte: 287,
-        endChar: 275,
-        startByte: 156,
-        startChar: 156,
+        endByte: 222,
+        endChar: 222,
+        startByte: 152,
+        startChar: 152,
       },
       contents:
-        '\n' +
-        '  <div\n' +
-        '    class={{local styles "container" "highlight"}}\n' +
-        '    data-test-field="😀😀a🎉🎉"\n' +
-        '  >\n' +
-        '    {{@user.name}}\n' +
-        '  </div>\n',
+        '\n    <div class={{styles.container}}>\n      Hello world!\n    </div>\n  ',
       endRange: {
-        endByte: 298,
-        endChar: 286,
-        startByte: 287,
-        startChar: 275,
+        endByte: 233,
+        endChar: 233,
+        startByte: 222,
+        startChar: 222,
       },
       range: {
-        endByte: 298,
-        endChar: 286,
-        startByte: 146,
-        startChar: 146,
+        endByte: 233,
+        endChar: 233,
+        startByte: 142,
+        startChar: 142,
       },
       startRange: {
-        endByte: 156,
-        endChar: 156,
-        startByte: 146,
-        startChar: 146,
-      },
-      tagName: 'template',
-      type: 'expression',
-    },
-    {
-      contentRange: {
-        endByte: 608,
-        endChar: 578,
-        startByte: 423,
-        startChar: 411,
-      },
-      contents:
-        '\n' +
-        '    <UserName @user={{@user}} />\n' +
-        '\n' +
-        '    <div class={{styles.container}} data-test-field="timestamp">\n' +
-        '      {{this.timestamp}}\n' +
-        '\n' +
-        '  <p> 😀😀😀 Hello! 🎉🎉🎉\n' +
-        '  </p>\n' +
-        '    </div>\n' +
-        '  ',
-      endRange: {
-        endByte: 619,
-        endChar: 589,
-        startByte: 608,
-        startChar: 578,
-      },
-      range: {
-        endByte: 619,
-        endChar: 589,
-        startByte: 413,
-        startChar: 401,
-      },
-      startRange: {
-        endByte: 423,
-        endChar: 411,
-        startByte: 413,
-        startChar: 401,
+        endByte: 152,
+        endChar: 152,
+        startByte: 142,
+        startChar: 142,
       },
       tagName: 'template',
       type: 'class-member',
