+++ packages/ast/template-tag/tests/replace-template/class-2.test.ts
@@ -5,43 +5,25 @@ import { replaceTemplate } from '../../src/index.js';
 test('replace-template > class (2)', function () {
   const oldFile = [
     `import Component from '@glimmer/component';`,
-    `import { local } from 'embroider-css-modules';`,
     ``,
-    `import styles from './example.css';`,
-    ``,
-    `const UserName = <template>`,
-    `  <div`,
-    `    class={{local styles "container" "highlight"}}`,
-    `    data-test-field="😀😀a🎉🎉"`,
-    `  >`,
-    `    {{@user.name}}`,
-    `  </div>`,
-    `</template>;`,
+    `import styles from './my-component.css';`,
     ``,
     `export default class MyComponent extends Component {`,
-    `  get timestamp(): string {`,
-    `    return 'yesterday';`,
-    `  }`,
-    ``,
     `  <template>`,
-    `    <UserName @user={{@user}} />`,
-    ``,
-    `    <div class={{styles.container}} data-test-field="timestamp">`,
-    `      {{this.timestamp}}`,
-    ``,
-    `  <p> 😀😀😀 Hello! 🎉🎉🎉`,
-    `  </p>`,
+    `    <div class={{styles.container}}>`,
+    `      Hello world!`,
     `    </div>`,
     `  </template>`,
     `}`,
+    ``,
   ].join('\n');
 
   const newFile = replaceTemplate(oldFile, {
     range: {
-      endByte: 619,
-      endChar: 589,
-      startByte: 413,
-      startChar: 401,
+      endByte: 233,
+      endChar: 233,
+      startByte: 142,
+      startChar: 142,
     },
     template: '\n    New contents\n  ',
   });
@@ -50,28 +32,15 @@ test('replace-template > class (2)', function () {
     newFile,
     [
       `import Component from '@glimmer/component';`,
-      `import { local } from 'embroider-css-modules';`,
       ``,
-      `import styles from './example.css';`,
-      ``,
-      `const UserName = <template>`,
-      `  <div`,
-      `    class={{local styles "container" "highlight"}}`,
-      `    data-test-field="😀😀a🎉🎉"`,
-      `  >`,
-      `    {{@user.name}}`,
-      `  </div>`,
-      `</template>;`,
+      `import styles from './my-component.css';`,
       ``,
       `export default class MyComponent extends Component {`,
-      `  get timestamp(): string {`,
-      `    return 'yesterday';`,
-      `  }`,
-      ``,
       `  <template>`,
       `    New contents`,
       `  </template>`,
       `}`,
+      ``,
     ].join('\n'),
   );
 });
