--- packages/ast/template-tag/tests/update-javascript/class-1.test.ts
@@ -1,43 +1,0 @@
-import { assert, test } from '@codemod-utils/tests';
-
-import { updateJavaScript } from '../../src/index.js';
-import { data, renameGetters } from '../helpers/update-javascript.js';
-
-test('update-javascript > class (1)', function () {
-  const oldFile = [
-    `import Component from '@glimmer/component';`,
-    ``,
-    `import styles from './my-component.css';`,
-    ``,
-    `export default class MyComponent extends Component {`,
-    `  <template>`,
-    `    <div class={{styles.container}}>`,
-    `      Hello world!`,
-    `    </div>`,
-    `  </template>`,
-    `}`,
-    ``,
-  ].join('\n');
-
-  const newFile = updateJavaScript(oldFile, (code) => {
-    return renameGetters(code, data);
-  });
-
-  assert.strictEqual(
-    newFile,
-    [
-      `import Component from '@glimmer/component';`,
-      ``,
-      `import styles from './my-component.css';`,
-      ``,
-      `export default class MyComponent extends Component {`,
-      `  <template>`,
-      `    <div class={{styles.container}}>`,
-      `      Hello world!`,
-      `    </div>`,
-      `  </template>`,
-      `}`,
-      ``,
-    ].join('\n'),
-  );
-});
