+++ packages/ast/template-tag/tests/update-javascript/class-2.test.ts
@@ -6,35 +6,17 @@ import { data, renameGetters } from '../helpers/update-javascript.js';
 test('update-javascript > class (2)', function () {
   const oldFile = [
     `import Component from '@glimmer/component';`,
-    `import { local } from 'embroider-css-modules';`,
     ``,
-    `import styles from './example.css';`,
-    ``,
-    `const UserName = <template>`,
-    `  <div`,
-    `    class={{local styles "container" "highlight"}}`,
-    `    data-test-field="😀😀a🎉🎉"`,
-    `  >`,
-    `    {{@user.name}}`,
-    `  </div>`,
-    `</template>;`,
+    `import styles from './my-component.css';`,
     ``,
     `export default class MyComponent extends Component {`,
-    `  get timestamp(): string {`,
-    `    return 'yesterday';`,
-    `  }`,
-    ``,
     `  <template>`,
-    `    <UserName @user={{@user}} />`,
-    ``,
-    `    <div class={{styles.container}} data-test-field="timestamp">`,
-    `      {{this.timestamp}}`,
-    ``,
-    `  <p> 😀😀😀 Hello! 🎉🎉🎉`,
-    `  </p>`,
+    `    <div class={{styles.container}}>`,
+    `      Hello world!`,
     `    </div>`,
     `  </template>`,
     `}`,
+    ``,
   ].join('\n');
 
   const newFile = updateJavaScript(oldFile, (code) => {
@@ -45,36 +27,17 @@ test('update-javascript > class (2)', function () {
     newFile,
     [
       `import Component from '@glimmer/component';`,
-      `import { local } from 'embroider-css-modules';`,
       ``,
-      `import styles from './example.css';`,
-      ``,
-      `const UserName = <template>`,
-      `  <div`,
-      `    class={{local styles "container" "highlight"}}`,
-      `    data-test-field="😀😀a🎉🎉"`,
-      `  >`,
-      `    {{@user.name}}`,
-      `  </div>`,
-      `</template>;`,
+      `import styles from './my-component.css';`,
       ``,
       `export default class MyComponent extends Component {`,
-      `  // Assigned new name`,
-      `  get __timestamp(): string {`,
-      `    return 'yesterday';`,
-      `  }`,
-      ``,
-      `    <template>`,
-      `      <UserName @user={{@user}} />`,
-      ``,
-      `      <div class={{styles.container}} data-test-field="timestamp">`,
-      `        {{this.timestamp}}`,
-      ``,
-      `    <p> 😀😀😀 Hello! 🎉🎉🎉`,
-      `    </p>`,
-      `      </div>`,
-      `    </template>`,
+      `  <template>`,
+      `    <div class={{styles.container}}>`,
+      `      Hello world!`,
+      `    </div>`,
+      `  </template>`,
       `}`,
+      ``,
     ].join('\n'),
   );
 });
