+++ packages/ast/template-tag/tests/update-javascript/template-only-3.test.ts
@@ -1,13 +1,13 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { updateTemplates } from '../../src/index.js';
-import { removeClassAttribute } from '../helpers/update-templates.js';
+import { updateJavaScript } from '../../src/index.js';
+import { data, renameGetters } from '../helpers/update-javascript.js';
 
-test('update-templates > template-only (2)', function () {
+test('update-javascript > template-only (3)', function () {
   const oldFile = [
     `import styles from './styles.css';`,
     ``,
-    `<template>`,
+    `export default <template>`,
     `  <div class={{styles.container}}>`,
     `    Hello world!`,
     `  </div>`,
@@ -15,15 +15,17 @@ test('update-templates > template-only (2)', function () {
     ``,
   ].join('\n');
 
-  const newFile = updateTemplates(oldFile, removeClassAttribute);
+  const newFile = updateJavaScript(oldFile, (code) => {
+    return renameGetters(code, data);
+  });
 
   assert.strictEqual(
     newFile,
     [
       `import styles from './styles.css';`,
       ``,
-      `<template>`,
-      `  <div>`,
+      `export default <template>`,
+      `  <div class={{styles.container}}>`,
       `    Hello world!`,
       `  </div>`,
       `</template>;`,
