--- packages/ast/template-tag/tests/update-templates/base-case.test.ts
@@ -1,12 +1,0 @@
-import { assert, test } from '@codemod-utils/tests';
-
-import { updateTemplates } from '../../src/index.js';
-import { removeClassAttribute } from '../helpers/update-templates.js';
-
-test('update-templates > base case', function () {
-  const oldFile = [`<template></template>;`, ``].join('\n');
-
-  const newFile = updateTemplates(oldFile, removeClassAttribute);
-
-  assert.strictEqual(newFile, [`<template></template>;`, ``].join('\n'));
-});
