--- packages/ast/template-tag/tests/update-templates/class-1.test.ts
@@ -1,41 +1,0 @@
-import { assert, test } from '@codemod-utils/tests';
-
-import { updateTemplates } from '../../src/index.js';
-import { removeClassAttribute } from '../helpers/update-templates.js';
-
-test('update-templates > class (1)', function () {
-  const oldFile = [
-    `import Component from '@glimmer/component';`,
-    ``,
-    `import styles from './my-component.css';`,
-    ``,
-    `export default class MyComponent extends Component {`,
-    `  <template>`,
-    `    <div class={{styles.container}}>`,
-    `      Hello world!`,
-    `    </div>`,
-    `  </template>`,
-    `}`,
-    ``,
-  ].join('\n');
-
-  const newFile = updateTemplates(oldFile, removeClassAttribute);
-
-  assert.strictEqual(
-    newFile,
-    [
-      `import Component from '@glimmer/component';`,
-      ``,
-      `import styles from './my-component.css';`,
-      ``,
-      `export default class MyComponent extends Component {`,
-      `  <template>`,
-      `    <div>`,
-      `      Hello world!`,
-      `    </div>`,
-      `  </template>`,
-      `}`,
-      ``,
-    ].join('\n'),
-  );
-});
