+++ packages/ast/template-tag/tests/update-templates/class-3.test.ts
@@ -1,0 +1,76 @@
+import { assert, test } from '@codemod-utils/tests';
+
+import { updateTemplates } from '../../src/index.js';
+import { removeClassAttribute } from '../helpers/update-templates.js';
+
+test('update-templates > class (3)', function () {
+  const oldFile = [
+    `import Component from '@glimmer/component';`,
+    `import { local } from 'embroider-css-modules';`,
+    ``,
+    `import styles from './example.css';`,
+    ``,
+    `const UserName = <template>`,
+    `  <div`,
+    `    class={{local styles "container" "highlight"}}`,
+    `    data-test-field="😀😀a🎉🎉"`,
+    `  >`,
+    `    {{@user.name}}`,
+    `  </div>`,
+    `</template>;`,
+    ``,
+    `export default class MyComponent extends Component {`,
+    `  get timestamp(): string {`,
+    `    return 'yesterday';`,
+    `  }`,
+    ``,
+    `  <template>`,
+    `    <UserName @user={{@user}} />`,
+    ``,
+    `    <div class={{styles.container}} data-test-field="timestamp">`,
+    `      {{this.timestamp}}`,
+    ``,
+    `  <p> 😀😀😀 Hello! 🎉🎉🎉`,
+    `  </p>`,
+    `    </div>`,
+    `  </template>`,
+    `}`,
+  ].join('\n');
+
+  const newFile = updateTemplates(oldFile, removeClassAttribute);
+
+  assert.strictEqual(
+    newFile,
+    [
+      `import Component from '@glimmer/component';`,
+      `import { local } from 'embroider-css-modules';`,
+      ``,
+      `import styles from './example.css';`,
+      ``,
+      `const UserName = <template>`,
+      `  <div`,
+      `    data-test-field="😀😀a🎉🎉"`,
+      `  >`,
+      `    {{@user.name}}`,
+      `  </div>`,
+      `</template>;`,
+      ``,
+      `export default class MyComponent extends Component {`,
+      `  get timestamp(): string {`,
+      `    return 'yesterday';`,
+      `  }`,
+      ``,
+      `  <template>`,
+      `    <UserName @user={{@user}} />`,
+      ``,
+      `    <div data-test-field="timestamp">`,
+      `      {{this.timestamp}}`,
+      ``,
+      `  <p> 😀😀😀 Hello! 🎉🎉🎉`,
+      `  </p>`,
+      `    </div>`,
+      `  </template>`,
+      `}`,
+    ].join('\n'),
+  );
+});
