+++ packages/ast/template-tag/tests/update-templates/template-only-1.test.ts
@@ -4,30 +4,9 @@ import { updateTemplates } from '../../src/index.js';
 import { removeClassAttribute } from '../helpers/update-templates.js';
 
 test('update-templates > template-only (1)', function () {
-  const oldFile = [
-    `import styles from './styles.css';`,
-    ``,
-    `<template>`,
-    `  <div class={{styles.container}}>`,
-    `    Hello world!`,
-    `  </div>`,
-    `</template>;`,
-    ``,
-  ].join('\n');
+  const oldFile = [`<template></template>;`, ``].join('\n');
 
   const newFile = updateTemplates(oldFile, removeClassAttribute);
 
-  assert.strictEqual(
-    newFile,
-    [
-      `import styles from './styles.css';`,
-      ``,
-      `<template>`,
-      `  <div>`,
-      `    Hello world!`,
-      `  </div>`,
-      `</template>;`,
-      ``,
-    ].join('\n'),
-  );
+  assert.strictEqual(newFile, [`<template></template>;`, ``].join('\n'));
 });
